requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
groq==0.4.2
httpx==0.27.2
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.1.0.post1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9